Feat: Added a new cron-clone.sh script and edited the README.md file in cloud-chatops
The cron-clone.sh script can be used to automatically pull docker-compose images via the cron.hourly, cron.daily, cron.weekly, or cron.monthly directories. The README files includes how to configure this.

ENH: Changed the cron-clone.sh file to not delete repo

Changed the script to fetch and clean instead of deleting everything. Also added error logging.

ENH: Edited the cron-clone.sh to cd in the correct order

The script now changes directory before it git fetches, cleans and resets.

MAINT: Edited the setup instructions and the .sh file

The instructions direct the user to change the permissions then copy the file. Changed the file name as not to include the file extension and removed o -pipefall as it wasn't permitted.