BUG: Function call was passing a string not a user object
The function run_personal_reminders takes a list of User objects but a string was given. To fix this we find the User object we need from the get_config function and then pass the user object in. Also, added the option to send a message to the user saying no PRs were found. Otherwise they would not get any message at all which is confusing