MAINT: Changing repo dictionary to a list
read_config.py will now return a list of repositories with string values such as "stfc/cloud-docker-images" instead of a dictionary such as {"stfc": ["cloud-docker-images"]}. This makes the finding PRs code simpler as it removes one iteration loop. The find prs modules should only be concerned with making requests and not formatting data. The read_config.py module will format the yaml dictionary into a list so the user is still using dictionaries as they visibly look easier to understand. Updated the tests to reflect this.