ENH: Add a custom route for weekly reminder messages
Added a second route that listens for authorised requests to send reminders to the workspace. Changing this from scheduling our own reminders each week to waiting for a request. This way when running multiple pods we don't have the number of pods amount of reminders. Updated test to reflect this change