ENH: Add slash_mrs to main.py
Adding the slash_mrs function as a copy of slash_prs. This will allow users to run the slash command and find open MRs. Removed the find_host command as this is not supported or needed in production. Also, removed docstrings from these functions as they are a duplicate of the slash_commands.py module