DOC: Remove deployments directory
Removing the deployments directory as it is bloated and unclear.
Dropping support for multiple methods of deployment as they are not used
and can be figured out by the user if they don't want to use the default
setup.
Also, moved all the template and config files into the INSTALL.md so we only have one copy of them