DOC: Updating the INSTALL.md instructions
Removing the deployments directory as it is bloated and unclear.
Dropping support for multiple methods of deployment as they are not used
and can be figured out by the user if they don't want to use the default
setup. For example, the current ChatOps deployment used for production
is much more different.