MAINT: Rename directory
This folder should contain modules that send messages to different platforms. It is unlikely we will need more than one module to send messages to slack. So we can rename this folder to notify and the module to slack. This makes imports clearer as you are importing notify.slack and if we used microsoft  teams we could have notify.teams etc.