ENH: Update pre-commit hooks for chatops
Adds pylint and pytest as hooks for the chatops directory. This makes
sure that each commit contains working, clean code. I could not add
pylint and pytest globally for the repo as each project has it's own
rc files for config.

I had to add both "cloud-chatops/chatops" and "chatops" to the path in
the pylint hook in case pylint is run from the repo root or project
root.